requirements.txt
langchain
sentence_transformers
torch
pypdf
qdrant-client
dotenv
chainlit